+++ CHANGELOG.md
@@ -17,7 +17,7 @@ and this project adheres to
 - 🐛(dimail) improve handling of dimail errors on failed mailbox creation #377
 -  💬(frontend) fix group member removal text #382
 -  💬(frontend) fix add mail domain text #382
-- 🐛(frontend) fix keyboard navigation on language picker #379
+- 🐛(frontend) fix keyboard navigation #379
 
 ## [1.0.2] - 2024-08-30